requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & AI
anthropic==0.21.0
openai==1.3.0
ollama==0.1.5

# Vector & Memory
chromadb==0.4.24
numpy==1.24.0
scipy==1.11.0

# GitHub Integration
PyGithub==2.1.1
requests==2.31.0

# Data Processing
pandas==2.1.0
markdown==3.5.0
PyYAML==6.0

# Storage & Serialization
sqlalchemy==2.0.23
sqlite3

# Async & Concurrency
aiohttp==3.9.0
asyncio-contextmanager==1.0.0

# Social Media (Future)
bluesky-sdk==0.0.26
lens-sdk==0.0.1

# Development
pytest==7.4.0
black==23.11.0
ruff==0.1.7
mypy==1.7.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
